requirements.txt
async-generator==1.10
attrs==23.1.0
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
exceptiongroup==1.1.1
ffmpeg-python==0.2.0
future==0.18.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
greenlet==2.0.1
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
joblib==1.2.0
moviepy==1.0.3
mutagen==1.46.0
nltk==3.8.1
numpy==1.24.3
outcome==1.2.0
Pillow==9.5.0
playwright==1.32.1
praw==7.7.0
prawcore==2.3.0
proglog==0.1.10
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.17
pyee==9.0.4
PySocks==1.7.1
regex==2023.3.23
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
textblob==0.17.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
websockets==11.0.2
wsproto==1.2.0
yt-dlp==2023.3.4
